Builds passing for mac. Proceeding to check on linux now
